BackWard implementation, we need to connect the .h files

+++ include/mnist_loader.h
@@ -6,7 +6,7 @@
 #include <cstdint>
 
 void loadMNISTImages(const std::string &filename, std::vector<std::vector<float>> &images);
-void loadMNISTLabels(const std::string &filename, std::vector<uint8_t> &labels);
+void loadMNISTLabels(const std::string &filename, std::vector<int> &labels);
 
 
 #endif // MNIST_LOADER_H
